updated examples of how to add extra functionality to Projog using Java #106

+++ src/core-tests/org/projog/example/CalculatableExample.java
@@ -17,6 +17,6 @@ public class CalculatableExample implements Calculatable {
    public Numeric calculate(Term... args) {
       Numeric input = castToNumeric(args[0]);
       long rounded = Math.round(input.getDouble());
-      return new IntegerNumber((int) rounded);
+      return new IntegerNumber(rounded);
    }
 }
